02_quant_research/01_factor_modeling_pca_eigenportfolios/requirements.txt
# Systematic Factor Modeling - Requirements
# Production-grade quantitative research project

# Core numerical computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine learning and statistics
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Data acquisition
yfinance>=0.2.28

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Optional: Enhanced functionality
# jupyter>=1.0.0  # For notebook analysis
# openpyxl>=3.1.0  # For Excel export
# tables>=3.8.0  # For HDF5 storage
